refactor: created VulkanMesh class

+++ include/Core/Consts.hpp
@@ -20,7 +20,7 @@ namespace Consts {
         {{0.5f, 0.5f}, {0.0f, 0.0f, 1.0f}},
         {{-0.5f, 0.5f}, {1.0f, 1.0f, 1.0f}}
     };
-    inline const std::vector<uint16_t> INDICES = {
+    inline const std::vector<uint32_t> INDICES = {
         0, 1, 2, 2, 3, 0
     };
 }
